vlserver: Avoid ubik abort for GetEntryBy* NOENTs
Currently, when the vlserver cannot find the given volume name or id,
VL_GetEntryByName* and VL_GetEntryByID* fail with VL_NOENT and abort
the ubik transaction. Aborting the ubik transacton means, among other
things, that we clear the application cache (by blanking
cachedVersion).

This means that any time a user looks up a nonexistent volume, we will
clear the application cache. When the vlserver services the next RPC,
it will re-populate its application cache, requiring the vlserver to
read various header information again (UpdateCache). This doesn't
cause any errors, but causes a significant number of completely
unnecessary ubik reads if there are many requests for nonexistent
volumes.

To avoid this, skip the ubik_AbortTrans call in these cases, and
instead just do a normal ubik_EndTrans (while still returning an error
from the Rx RPC). This commit only changes the VL_NOENT codepath for
GetEntryBy* calls, since they are by far the most common source of
VL_NOENTs in practice. There are other codepaths that could maybe also
be changed in the same way, but this commit attempts to limit the
scope of changes, at least for now.

Issue found by (and commit written with the collaboration with)
mvitale@sinenomine.net.

Change-Id: Ic13f7899763de6d0170c7aa922981f7e47f2808a