vlserver: Introduce vl_EndTrans/vl_AbortTrans
Instead of calling ubik_EndTrans(ctx->trans), convert all callers to
go through a new function, vl_EndTrans(ctx). Currently, the only new
behavior this adds is that we clear ctx->trans, which ensures we don't
try to access our transaction after ending or aborting it. But this
allows us to easily change the code around EndTrans in future commits
in one place.

Do the same for ubik_AbortTrans, as well.

Change-Id: Id14b0cec4664b16a818d3bb30da599965b6c5f1c