rx: Introduce rxi_WaitForTransmitWindow
Currently, rxi_WriteProc and rxi_WritevProc contain the same chunk of
code to wait for the transmit window on the rx call to advance past a
certain point. Consolidate this code out into a common function,
called rxi_WaitForTransmitWindow.

This commit should have no effects; it is just code reorganization.
Technically the check for 'if (call->error)' in rxi_WriteProc is
effectively moved to outside of the waiting while() loop, but the loop
is broken when call->error is nonzero anyway, so this should have no
effect.

Also introduce rxi_TransmitQueueIsTooFull(), to further simplify the
conditionals in here and try to make the code more clear.

Change-Id: I02f37c7cbbd45847d1b42d03335fa6166e7a1a20