rx: Introduce rxi_Read/ReceivePacketList
Currently, the rx listener processes packets like the following
pseudocode:

    for (;;) {
        rxi_ReadPacket(sock, packet, &otherdata);
        rxi_ReceivePacket(sock, packet, &otherdata);
    }

That is, we read a single packet off of the wire, and then process
(aka "receive") it.

When processing a packet associated with a specific rx call, we need
to lookup the connection and call, and lock various things in the
process. If we receive multiple packets for the same call
simultaneously, then it would be more efficient to lock the relevant
call once, and then process all of those packets before unlocking the
call. This commit introduces some new functions to allow for that, to
read and process multiple rx packets at once.

Thus, with this commit, the rx listener code looks more like this:

    for (;;) {
        rxi_ReadPacketList(sock, &packetlist, &otherdata);
        rxi_ReceivePacketList(sock, &packetlist, &otherdata);
    }

We receive at most rxi_RecvAtOnce packets at a time, and this commit
sets rxi_RecvAtOnce to 1. So this commit by itself does not improve
the performance of receiving packets by the listener, but future
commits will do so.

Change-Id: I8531427caad26de3e38abc7b1aaf6cf528fa9395