rx: Introduce 'flags' argument to rxi_SendList
Currently, the last argument to rxi_SendList is just a boolean, which
indicates whether or not the RX_MORE_PACKETS packet header flag should
be set. To make it easier to add more simple boolean switches to
rxi_SendList, change this last argument to be a 'flags' bitfield, and
define a flag, RXI_SENDLIST_MORE, to be identical to the current
behavior.

This commit doesn't define any new flags or change the behavior of the
'more' flag, so this commit should not have any visible impact. Future
commits will add different flags.

Change-Id: I4f18b7caedf7cd09c9c151ad6f046a86703d499f