rx: Introduce rxi_SendPacketDgrams
Currently, rxi_SendList calls one of two functions to actually send
the underlying packet data:

- rxi_SendPacket, to send a single packet

- rxi_SendPacketList, to send multiple Rx packets inside a single UDP
  packet (an Rx jumbogram)

Introduce a new variant, called rxi_SendPacketDgrams, which sends a
list of packets, but as separate UDP packets instead of sending them
all in the same packet. This allows callers to send a list of packets
"in bulk", avoiding some unnecessary locking/unlocking, but still
avoiding large jumbogram packets.

Allow callers of rxi_SendList to use this new variant when a new flag,
RXI_SENDLIST_DGRAMS, is given, and make rxi_SendXmitList use this flag
when we know we're not going to send any jumbograms. This also allows
rxi_SendXmitList to skip processing each packet in the given list,
since we're going to just send the whole list effectively at the same
time.

This commit also introduces a new undefined struct,
rxi_xmit_dgramlist, as a parameter to rxi_SendPacketDgrams. For now,
this is always NULL, but will be used in future commits.

Change-Id: Iea84466a407da7cab333ac0dc3c938be013571e1