rx: Defer rxi_Start calls during rxi_WriteProc
Currently, when an application is sending a lot of data to peers,
rxi_WriteProc/rxi_WritevProc tends to add many packets to the transmit
queue (more than our actual window size), and then waits for the
transmit queue to shrink as the peer ACKs our packets. The listener
handles running rxi_Start when we need to transmit data to the peer,
once the peer has ACKed data to advance the window. This allows the
application thread to sleep while the listener takes care of the
simple packet transmissions.

However, when an application is sending a lot of data from many
different threads, the listener can become a bottleneck, because it is
handling most of the rxi_Start calls (and most of the sendmsg()
syscalls) for many different Rx calls in many different threads.

To alleviate this bottleneck, we can make it so some of the rxi_Start
calls are made from the rxi_WriteProc/rxi_WritevProc thread instead of
from the listener. Specifically, while rxi_WriteProc/rxi_WritevProc is
running, listener calls to rxi_Start instead just set a flag on the
call, RX_CALL_NEEDS_RXISTART, and possibly wakeup a thread that's
sleeping in rxi_WriteProc/rxi_WritevProc waiting for the window to
advance.

To do this, rxi_WriteProc/rxi_WritevProc set a flag on the call
(RX_CALL_DEFER_RXISTART) before they sleep for window, and clear that
flag before they return. If the listener tries to call rxi_Start and
it sees that flag set, we skip the call to rxi_Start and just set
RX_CALL_NEEDS_RXISTART instead (rxi_StartDefer). The code in
rxi_WriteProc/rxi_WritevProc checks if RX_CALL_NEEDS_RXISTART is set
when woken up, and calls rxi_Start on the call if so
(rxi_StartIfDeferred).

If we are running with LWP, avoid all of this "deferred" logic, and
just make direct calls to rxi_Start in the listener like normal. LWP
is all single-threaded regardless, so it gains no benefit from
shifting load to another thread.

Change-Id: I7c1186fc0372d30497d3bac94a11a8ee7d96225b