rx: Use SO_REUSEPORT for multiple listener threads
When we are using multiple listener threads, create a separate socket
with the SO_REUSEPORT flag set for each listener thread. This allows
the OS to more evenly spread packet receives between threads, and
makes it more likely for us to receive packets for the same call in
the same thread (OSes tend to determine which socket to send a packet
to by hashing the ip and port of the src and dest by default).

To do this, we add a new argument to rxi_BindSocket, 'reuseport',
which is just a flag to say whether or not to use SO_REUSEPORT on the
socket. For the initial bind, we first try to bind the socket without
SO_REUSEPORT to make sure we get an error if someone else is trying to
use the same socket. We then close the socket and do the bind again to
get the functionality of SO_REUSEPORT.

SO_REUSEPORT is available on at least Linux 3.9+ and Solaris 11.4+.

Change-Id: Ie4d96d11ef6fd961cbf39c79d484c39227ad2fc8