rx: Define conn->flags as a uint32
The 'flags' field in struct rx_connection is only defined as a u_char
currently, but we define and use connection flags up to 256
(RX_CONN_NAT_PING). A u_char can only go up to 255, so there is no way
this flag is working correctly right now.

Expand 'flags' to be an afs_uint32 (like the flags in struct rx_call),
so we can at least store all of our relevant flags.

Change-Id: I93c3b44751e0bdc87263e41067c313fccd4c080e