Introduce ViceLog_limit(), afs_warn_limit()
Introduce a new variant to ViceLog(), called ViceLog_limit(), that
implements a primitive form of rate-limiting. This can be useful when
we want to log some information about an unusual event, but that event
might happen very frequently, and so just using ViceLog() may result
in flooding logs and filling disks.

To keep things simple, these new ViceLog variants avoid using locks or
atomics to guarantee that the rate of log messages stays within the
given limit. Instead, we just treat the limiting as best-effort, and we
may exceed the limit temporarily if we race. But this shouldn't matter
much, and should be good enough to avoid filling disks with log messages
(e.g., it may be possible to see an extra 5 log messages per time
interval, but we're more concerned about avoiding logging 5 million
messages).

Also introduce afs_warn_limit(), which is the same thing for
afs_warn() for use in the kernel.

Change-Id: I659276d0bf9979fcbb7b5d11dfc543cb301c6027