[from ubik-rdr branch] ubik: Make dbFlags protected by just DBHOLD
Currently, dbFlags is one of the fields protected by DBHOLD and
UBIK_VERSION_LOCK; reading from it requires either lock, and writing
to it requires both. Usually fields in ubik_dbase are protected just
by DBHOLD, but one place in beacon.c needs to check dbFlags (to see if
DBWRITING is set) without acquiring DBHOLD. Thus, the
UBIK_VERSION_LOCK protection was added to allow for beacon.c to read
the value without DBHOLD.

The current scheme makes it difficult to make use of new database
flags, since we need to acquire two locks to set or clear any flags.
So to make this easier for future commits, make dbFlags just protected
by DBHOLD, and not by UBIK_VERSION_LOCK.

To solve the issue of accessing dbFlags in beacon.c, add a new field
'db_writing' in version_globals, which is just protected by
UBIK_VERSION_LOCK, and is set whenever dbFlags&DBWRITING is set, and
cleared whenever dbFlags&DBWRITING is cleared.

Change-Id: I70bcccf786e622ee4041f699b75f2201cd477366