rxgk: Add and use 524 conversion logic
Introduce rxgk_524_conv_id(), a function to convert a krb5-backed GSS
identity into a krb4-based identity (using the same rules as rxkad to
convert krb5 princs into krb4 princs). Use this new function in the
ptserver, to lookup the id of a caller using the existing krb4-based ids
in the ptdb.

Change-Id: I455942d73824e14d09082b285137bdc58208a4b5