viced: Introduce -s2scrypt
Currently, the fileserver always communicates with the ptserver and
vlserver with rxkad. Add the -s2scrypt option to the fileserver (like
exists for volserver and vlserver), which can specify whether to use
rxgk ('-s2scrypt rxgk-crypt') or rxkad ('-s2scrypt always'). When using
'-s2scrypt rxgk-crypt' and the vlserver supports rxgk, the vlserver will
record that the fileserver supports rxgk, and will advertise to clients
that the fileserver supports rgxk.

Change-Id: I3905a9822261c085282e9539a3f1cf3d43e41265