rxgk: Implement single-token AFSCombineTokens()
Implement AFSCombineTokens() for real in the vlserver, detecting if the
given fileserver supports rxgk, and returning a fileserver-specific
combined token for that fileserver. Only support AFSCombineTokens() for
single-token calls for now; we don't yet support combining user tokens
with per-CM tokens.

To detect whether a fileserver supports rxgk via the cell-wide key,
define a new per-MH flag in the vldb, VL_MHFLAG_RXGK_CELLWIDE (0x1). Set
this flag when the fileserver registers its addrs over an rxgk
connection. Also clear the flag if the fileserver registers over an
rxkad connection (and log a warning); the rxgk spec specifically says to
not do this to avoid downgrade attacks, but we do this for now while
rxgk support is still experimental.

Create a new variant of rxgk_make_token(), called
rxgk_make_token_printedok(), to allow us to make a single-token
combined token with printed tokens.

Define rxgk_afscombine1_keydata(), a variant of rxgk_afscombine1_key(),
to perform the key-combining operation on opaques and instead of
rxgk_key objects, for convenience.

Change-Id: Iebd7ed76c8ba03585af4eff64d211f91bb37c317