rxgk: Introduce rxgk_CombineSingleClientSecObj()
Introduce the function rxgk_CombineSingleClientSecObj(). This function
calls the single-token AFSCombineTokens() operation on a vlserver to get
a fileserver-specific client security object, which can be used to
communicate with a fileserver.

To implement this, also implement rxgk_afscombine1_key(), to derive the
combined key locally. Also change fill_authenticator() to include the
relevant UUIDs in the appdata in the rxgk response authenticator.

This commit doesn't add any callers of rxgk_CombineSingleClientSecObj();
future commits will do so to enable clients to talk to rxgk fileservers.

Change-Id: I1cc3905a520f02b8931348436b183430d078d45a