rxgk: Decl AFSCombineTokens' afsUUID as a pointer
The RPC-L for AFSCombineTokens was added in commit
ee56e925f0 (Add rxgk boilerplate), but the 'destination' arg
is supposed to be an afsUUID*, not an afsUUID. Declare it as a
pointer.

This doesn't change anything in the wire protocol (which can be
verified by checking the generated rxgk.ss.c et al files); it just
allows us to pass the 'destination' arg as a pointer in our C calls,
instead of passing around an entire struct afsUUID by value.

Change-Id: I6f0ae9a230400cc51df548966ecae2a876878891