afs: Pass explicit service to afs_ConnBy*
Currently, afs_ConnBySA infers what Rx service id to use for the
underlying Rx conn by looking at the requested port number. This
assumes (correctly) that any caller that wants to contact port 7003
wants to use service 52, and any caller that wants to contact port
7000 wants to use service 1. This works, but it makes it difficult for
the client to ever contact different services on the same port.

To make this more flexible for future changes, change afs_ConnBySA to
accept an additional argument, specifying the service id to use, and
remember the service id in our cached sa_conn_vector.

Adjust all callers to give either the RXAFS or VL service id
accordingly. Also add named defines for the RXAFS and VL service ids,
to make it clearer which service a caller is requesting.

Also remove the unused 'cell' argument in afs_ConnBySA, to try to
reduce the number of arguments.

Change-Id: I5df2d3015b9fbc5bb27fa6f06972b48f9d05de68