rxgk: Rename PrAuthName -> RXGK_PrAuthName
The struct PrAuthName is from
draft-brashear-afs3-pts-extended-names-09, which will be used in
ptint.xg for ptserver RPCs. We also use the struct in rxgk
(specifically, in the RXGK_Token struct). Using the exact same name
causes errors when, for example, rxgen generates
the xdr_PrAuthName() function in two different object files
(ptint.xdr.o and rxgk_int.xdr.o), amongst other annoyances.

Ideally we would use the same type in both xg files, and get rxgen to
consolidate the duplicate definitions, but figuring that out is tricky
or involves hacky workarounds. At least for now, just take the easy way
out and rename the copy in rxgk to RXGK_PrAuthName.

Change-Id: I7dc8be8eac9c3444519e1975b49f1f3dd78597c5