auth: Zero 'tbuffer' when getting tokens
In ktc_GetTokenEx and GetToken, we read our tokens from the kernel
into 'tbuffer'. Some memory-checking tools such as valgrind do not
notice that our pioctl has populated this buffer, causing valgrind to
flag many call sites as using uninitialized memory.

Zero this buffer before using it to more obviously guarantee that the
memory in this buffer is initialized to something. This quiets such
tools, but is also a generally good idea, to ensure the memory is set
to 0 if we do accidentally read more data than is filled by the
pioctl.

Change-Id: Ibcdb1917d080f88c7d39f5afa85987bb182596b9