[from rxgk branch] rxgen: Introduce xdrfree_type()
Currently, the most common way to free an object allocated by xdr is
to call:

    xdr_free((xdrproc_t) xdr_foo, &foo);

Which runs the given object through the xdr routines with the XDR_FREE
operation. This works, but is not typesafe; if the wrong xdr_foo
routine is given to xdr_free, we will silently run the wrong xdr
routines, potentially freeing corrupt memory, etc.

There is a typesafe way to free an xdr object, if the caller basically
copies the implementation of xdr_free():

    {
        XDR x;
        x.x_op = XDR_FREE;
        (void)xdr_foo(&x, &foo);
    }

But this is rather cumbersome, and so is uncommon.

To allow convenient freeing of xdr objects in a typesafe manner,
introduce a generated function for each xdr type, called
xdrfree_foo(). This just calls xdr_free(), but the types of its
arguments guarantee that the correct type is given to xdr_free().

Change-Id: I80c7f2d61c6445f5310a765cef4079550d3a5a88
(cherry picked from commit 48801efefc4922f3162406cecacf10a35480bc91)