bozo: Introduce bos -rxgk, -rxgk-cellwide
Update bos to negotiate rxgk credentials when communicating with the
bosserver when the new option -rxgk is given. By default, this involves
communicating with the RXGK_ rx service on the bosserver itself, instead
of talking to a vlserver (like we normally do to get a token with
gklog). If -localauth is given with -rxgk, by default we'll use
rxgk.keytab to negotiate rxgk creds, since we don't have access to the
bosserver's ephemeral in-memory key.

Also introduce the new option -rxgk-cellwide. When this is given with
-rxgk, we will use an rxgk token encrypted with the cell-wide key in
KeyFileExt, instead of directly negotiating credentials with the RXGK_
rx service with the bosserver-specific ephemeral key.

To help implement this new functionality, introduce the new function
rxgk_NegotiateClientSecObj(), as a variant for the existing
rxgk_NegotiateClientToken().

Document these new options.

Change-Id: I7d30387d88d9a4621e92512a95e3b05bc96e80b0