bozo: Implement rxgk bosserver integration
Update bosserver to provide the RXGK_ rx service, to allow GSS
negotiation to happen with an ephemeral randomly-generated
token-encrypting key (instead of the cell-wide rxgk key), so 'bos'
clients can work with non-localauth rxgk.

In this commit, we also have bosserver accept the cell-wide rxgk key if
it's available, which 'bos -localauth' currently uses, so -localauth can
work similarly to rxkad. This also allows -localauth creds to work when
the GSS acceptor identity is not available.

To accomplish this, use a random kvno for our ephemeral key, trying to
avoid kvnos for the rxgk cellwide key on disk. We give a different
getkey function to rxgk_NewServerSecurityObject() (fallback_getkey()),
which uses the ephemeral key if it matches; if it doesn't match, we call
the normal afsconf_GetRXGKKey() to use the cellwide key.

Change-Id: I62dfbb4c4e960023ed061817a0096c3fc7d49b38