rxgk: Interpret server-side AFS app data
Update the rxgk server routines to actually process the AFS-specific
appdata in the authenticator. If the peer provides a key for a secure
callback channel, reject the client (with RXGK_SEALED_INCON), so we
don't appear to agree to use the callback key.

Allow server daemons to provide a server uuid (via
rxgk_NewServerSecurityObject), and reject clients with
RXGK_SEALED_INCON if they provide the wrong uuid. Allow servers to
retrieve the client uuid for a conn, with rxgk_ServerGetPeerUUID.

Update the fileserver to use these new facilities, effectively
requiring clients to provide the correct uuid in the authenticator.
Use the client-provided client uuid in h_GetHost_r when available, to
reduce the number of required RXAFSCB_TellMeAboutYourself calls to
clients.

Change-Id: I67f9a53d6cb3513f7a1c17ac6cf142090dbec392