vlserver: Reduce unnecessary pointer casts
Many areas in the vlserver code cast various pointers to char* when
using them as an arbitrary buffer. This is annoying at best, or even
potentially error-prone since the cast may silence compiler warnings
we may otherwise care about.

The casts are unnecessary when going through a void*, so eliminate
casts that are assigning a void*. Change the signature of a few
functions to use a void* buffer as well, and eliminate the related
casts.

Change-Id: Ib336ee4e34f4b9f3625620887ffb654e7f00d821