vlserver: Make vlentryread/write use an nvlentry
Currently, vlentryread() and vlentrywrite() take a void* buffer and a
length, similar to read() and write(). But these functions always
interpret the given buffer as a struct nvlentry, returning an error if
the given length doesn't match sizeof(nvlentry).

Change these functions to just actually accept a struct nvlentry
instead, to make it a little clearer what is going on and avoid the
need to check the given length.

Change-Id: Id2f7c3d3525e69f119d22941354bfd1543821341