[from ubik-rdr branch] ubik: allow read transactions during recovery
Currently, it is not possible to perform read and write transactions on
sites that are receiving or sending a database to another peer. Any
transaction triggered during this phase is blocked by DBHOLD. Knowing
that the database should not be modified during this process, write
transactions have to be blocked and the current approach is correct.
Unfortunately, the same is not true for read transactions.

As mentioned before, read transactions are blocked by DBHOLD on sites
that are sending or receiving a database. In the first case, the
database is not modified and there is no reason to block reads. On the
other hand, if the site is receiving a database from another peer, the
database will be replaced and read transactions should not be allowed
during this process. However, the data received from other peers are not
written directly into the database. Instead, these data are stored in a
temporary file which will eventually replace the database at the end of
this stage. With that being said, the only moment where reads should be
blocked is when the database is replaced by this temporary file.

Allowing read transactions while a database is being transferred can
mitigate the side effects of several synchronizations triggered by other
issues (such as problems on the network). In this scenario, this patch
would allow read transactions and the whole cell would not be completely
blocked.

To make it possible, this commit introduces two new database flags
(DBSENDING, DBRECEIVING) used to allow reads but still block writes
while a site is sending or receiving a database. Also, add a few small
functions (ubik_*_db_flags) to help with waiting for and using these
flags.

Change-Id: Ic8ffa01811fda2fdb4a50fb7d58462e495ce880c