vlserver: Abort transaction on vlsetcache error
If vlsetcache() returns an error here, Init_VLdbase() will return an
error to its caller. And when Init_VLdbase() returns an error to its
caller, its callers do not end the transaction (since we suppsedly
failed to initialize the transaction). This leaks the transaction,
which also keeps the lock on the database held, blocking future writes
(and reads, if this was a write transaction).

Currently, the only error vlsetcache() can return is VL_NOMEM, and
only when we need to allocate memory in vlexcpy(), which is very rare.
But future commits may add new errors to this code paths that are more
common.

Change-Id: I6db2e7c21e0c86329b44b39ff5c11d83cfdf161e