vlserver: Introduce new vldb_upgrade utility
Add a new utility, called vldb_upgrade, that can convert between the
vldb4 and vldb4-kv formats for the VLDB.

The utility's general structure is intended to lend itself to be
somewhat easily adapted to handle newer VLDB formats (which may be
very different from vldb4). But it may need further reworking when
such formats are actually defined.

Change-Id: I2ac2efd1eb6b7a334e68657e22b8e6825856dc59