vldb_convert: Rename to vldb_convert_old
The current vldb_convert utility is only for converting between very
old formats of the VLDB (versions that predate OpenAFS 1.0). Change
its name to vldb_convert_old, to try to prevent users from thinking
that it can be used to convert the VLDB between modern formats (VLDB
version 4 and newer).

Change-Id: Icfa95e3cb200a3ed2f5669aa33edca4c9a34cbab