vldb_check: Check MHC for duplicate MH block
Currently, we set the flag 'MHC' in 'record[rindex].type' to indicate that a
record is an MH block, and we set 'FRC' to indicate a record is a
vlentry on the free chain.

This test inside vldb_check is clearly intended to check if a record
has already been identified as an MH block (that is, it's already been
referenced by one of the ex_contaddrs entries), but it's checking for
the FRC flag, not MHC. Check for MHC instead.

This mistake appears to be a simple copy/paste mistake, from copying a
very similar segment of code for checking the free chain (which does
use the FRC flag). This code has been this way since OpenAFS 1.0.

Change-Id: I9a79991425f7620642ad6561c16d382761fcb5cc