[from rxgk branch] rx: Introduce rx_opaque_stringify
Add a new function to easily output the contents of rx_opaque objects
(in hex). Add a test for the new function, and while we're doing so,
add some tests for the rest of the rx_opaque functions, as well as
some afstest_common helper functions to make it easier to test with
opaques.

Change-Id: Ib43b94046d1ab47f1370a77064e9a335ae3622b5