rx: Introduce rx_opaque_cmp(), _stringify()
Add a new function, rx_opaque_cmp, to simplify comparing two rx_opaque
objects. Also add a new function, rx_opaque_stringify, to easily
output the contents of rx_opaque objects (in hex).

Add tests for the new functions, and while we're doing so, add some
tests for the rest of the rx_opaque functions, as well as some
afstest_common helper functions to make it easier to test with
opaques.

Change-Id: Ib43b94046d1ab47f1370a77064e9a335ae3622b5