rxgen: Introduce "shared_struct" keyword
When rxgen generates code for an RPC-L struct (say 'ubik_version'), it
generates the 'struct ubik_version' definition in header file
(ubik_int.h), and the 'xdr_ubik_version' serialization function in a
separate C file (ubik_int.xdr.c, included in liboafs_ubik and
libafsauthent). If another .xg file defines an XDR type that contains
'ubik_version', using the resulting .xdr.c code is a bit cumbersome:

- If we provide our own xdr_ubik_version function, we cannot link with
  ubik (xdr_ubik_version will be a duplicate symbol)

- If we don't provide our own xdr_ubik_version function, we must
  always link with ubik, even if we're not using any types that
  reference the ubik_version struct

While it's possible to work around these limitations, it would be
simpler to just define xdr_ubik_version as a macro function in the
generated header file. That way, anyone who includes ubik_int.h can
use the ubik_version type, with the standard header include guards
preventing any duplicate definitions.

So to be able to share type definitions across subsystems in this way,
introduce a new keyword, "shared_struct". This behaves the same as the
regular "struct", but the xdr definitions for the resulting type go in
the generated header file (as macros) instead of the generated C file.
We only do this for structs declared with shared_struct, to retain the
current behavior for existing structs.

This commit just adds support to rxgen for the new keyword; later
commits will actually make use of it.

Change-Id: I3edb1491e24baa8e4590cca3a88648f9a2f0bf73