cmd: Allow "--" as an end-of-options delimiter
Currently, it is very cumbersome to give any libcmd option a value
that starts with a hyphen followed by a non-digit. For example, if you
wanted to lookup pts users with the names "-foo" and "-bar", you can't
just give those as the first arguments to "pts examine", since libcmd
interprets them as (nonexistent) option names:

    $ pts examine -foo -bar
    pts: Unrecognized or ambiguous switch '-foo'; type 'pts help examine' for detailed help

The only way to specify such a value is to give the option name for
each value (since commit a2f1ca5f "cmd: Support splitting switches and
values with '='"):

    $ pts examine -nameorid=-foo -nameorid=-bar
    pts: User or group doesn't exist so couldn't look up id for -foo
    pts: User or group doesn't exist so couldn't look up id for -bar

This is rather cumbersome to deal with in automated calls to our
commands, and especially if these are typed out by a user in an
interactive session.

To improve this, catch up to POSIX.2 from 1992 by interpreting "--" as
a delimiter for the end of options; anything after the "--" is
interpreted as an operand. This makes it possible to run a command
like the above example:

    $ pts examine -- -foo -bar
    pts: User or group doesn't exist so couldn't look up id for -foo
    pts: User or group doesn't exist so couldn't look up id for -bar

Since now specifying an option (or alias) name of "--" won't work
properly, prevent the creation of options with such a name. Add the
new error code CMD_BADNAME to handle this case.

Change-Id: I4c2762c26aa8f2076518462ce4d49f9c1be5913c