Use getprogname/setprogname more
A few of our tools (e.g. aklog, compile_et, pts) try to manually
calculate the current program name, such as by looking at argv[0], or
libcmd's a0name. The logic is not only duplicated, but isn't always
consistent (sometimes we calculate the basename, sometimes not), and
is a little silly to do manually, since we already have setprogname()
and getprogname() to do this.

In addition, the ugen_ClientInit* functions are already using
getprogname() to get the current program name, but not all callers
are using setprogname() to set the program name (e.g. vos). This means
that vos can easily show an unpleasant string instead of the program
name, for example:

    $ vos listvldb root.cell
    <unknown>: Could not get afs tokens, running unauthenticated.
    [...]

To improve this situation, more consistently call setprogname(argv[0])
and getprogname() near the start of main(), and remove the redundant
argv[0]-parsing logic in some tools and tests.

Change-Id: I4e5108b2fc39d5a5ad543e6923f99c59d7c20298