tests: Add prdb db format tests
Add some basic tests to check ptserver code that interprets the prdb
database format. That is, check that using a given prdb.DB0 results
in the correct data being returned from 'pts' commands.

To do this, adapt the ubiktest framework to accept different server
types to run against (the new type 'afstest_server_type'). Change
afstest_StartVLServer to afstest_StartServer, which accepts an
argument to say which type of server to start.

Change-Id: I33b65c76c722064f43a34794c5c4a7315a70c56b