ubik: Introduce ubik_InstallWriteHook
Currently, the ptserver redirects the ubik_dbase->write function
pointer when supergroups is turned in, in order to intercept writes to
the ptdb to invalidate entries in the 'sg_found' and 'sg_flagged'
caches.

This scheme is maybe questionable for a variety of reasons, but we can
at least fix the ptserver from reaching into ubik internals by
introducing a new function, ubik_InstallWriteHook. This way, the
ptserver can call a function to install this write hook, and so the
ptserver doesn't need to directly interact with ubik internals and
doesn't need to include ubik_internal.h.

Change-Id: Idb59e2eb883ba9a497f6ad2e4845b9400b4632d3