ubik: Introduce database freezing/dumping
Introduce some openafs-ctl subcommands to manage 'freezing' a ubik db,
as well as dumping a db from a running server.

For example, to dump the vldb from a running vlserver, the user can
run:

    $ openafs-ctl vldb-dump /tmp/foo.DB0

Which freezes the vldb, then copies the db data to the indicated path,
and unfreezes the vldb. While the database is frozen, the vlserver can
still serve read requests from clients, but the vlserver cannot write
to the database or sync the database with other sites via urecovery.

While the database is frozen, the freeze context is associated with a
running afsctl call. If that call ends (if, for example, the
'openafs-ctl' process unexpectedly dies), the freeze is immediately
aborted and the database is fully usable again.

A couple of additional commands are also introduced here, for more
complex management of freezes:

- 'openafs-ctl vldb-freeze-run', which allows running custom commands
  while the database is frozen

- 'openafs-ctl vldb-freeze-abort', which allows an administrator to
  explicitly abort a running freeze

This commit adds support for the vldb (via the vlserver) and the ptdb
(via the ptserver).

Credit to Marcus Watts for the original inspiration for this general
idea way back in 2010.

Change-Id: I836f79a2ae2d6ace3cd65293087c130949712081