ubik: Introduce online db restores via openafs-ctl
Introduce a few new commands to the 'openafs-ctl' suite, allowing a
user to restore a ubik dbase on a running server. For example:

    $ openafs-ctl vldb-restore /tmp/new.DB0 -backup-suffix .BAK

This uses the same 'freeze' mechanism as vldb-dump et al, and so can
also be used with vldb-freeze-run and vldb-freeze-abort. If the freeze
ends unsuccessfully, the database is reverted back to the original,
unless the db has already been distributed to other sites.

By default, the newly-installed database is immediately distributed to
the other sites in the quorum. But if 'vldb-restore -dist skip' is
given, we don't distribute the dbase immediately; the dbase can be
later distributed via vldb-freeze-dist.

This potentially allows an administrator to install a database to just
the local site, then perform some tests or sanity checks, and then
only distribute the installed db if those checks succeed. The newly
installed database could be a completely new database, or could be
derived from the existing VLDB obtained via 'vldb-dump' (it could be
converted to a new VLDB version, or modified/fixed offline in some
way, etc).

Change-Id: I436e9c2c4f85e50fb2f076c1ae1008c3a10f3101