ubik: Add callback to check installed db sanity
Add a new field to ubik_serverinit_opts (dbcheck_func), which, if
speicfied, is called when someone is installing a new database via
e.g. 'openafs-ctl vldb-restore'. If the function returns an error, the
database is treated as invalid, and the install is aborted. This can
help prevent situations where an administrator accidentally tries to
restore an invalid db (e.g. a completely blank database, random
garbage, or someone tries to restore a ptdb to a vlserver).

Implement and specify this function for the vlserver. We don't
implement this for the ptserver in this commit, just because the way
the ptserver code is structured makes it difficult (too many globals,
etc).

Change-Id: I8c5c24f02a892aee53eab7c88d1b4589b2d3cdf8