tests: Introduce afstest_StartServerOpts
Introduce a new variant of afstest_StartServer*, called
afstest_StartServerOpts, which allows callers to specify more options
when starting a test server. Specifically, this adds the ability to
pass extra command-line arguments to the server (via 'extra_argv').

Also add this functionality to struct ubiktest_ops and pass along the
new args appropriately.

Note that nothing in this commit actually specifies any new arguments;
this just implements the ability to do so.

Change-Id: I2c21584f94a4fcfc1a1b60d8b77b382b0c092791