ubik: Introduce DISK_GetFile2/DISK_SendFile2
Introduce and use a couple of new RPCs: DISK_GetFile2 and
DISK_SendFile2. These are variants of the existing DISK_GetFile and
DISK_SendFile calls, but allow for sending/receiving different types
of ubik databases (currently flat-file or KV).

The new RPCs are structured a little differently than the existing
GetFile/SendFile calls; instead of passing some data as RPC arguments
(like file length or ubik version), everything is in the database
"stream" sent over the call, which contains a header and footer
(ubik_dbstream_header and ubik_dbstream_footer). This helps the code
paths for sending and receiving whole databases be more similar,
whether client-side or server-side.

Change-Id: I37e11e7fb4fbfba7fc66cb463542d29c79738544