ctl: Introduce openafs-ctl
Introduce a new command suite for interacting with daemons over the
afsctl subsystem, called 'openafs-ctl'. Add a simple command for
querying ubik db info, called '*db-info' (vldb-info for the vldb,
ptdb-info for the ptdb). The actual code for this subcommand is in
src/ubik/ubik_ctl.c, which generates the relevant libcmd subcommands
for different ubik applications.

Add the server-side routines for ubik as well, adding the necessary
hooks to the vlserver and ptserver.

Change-Id: I7b22a0fb6338c5010d31b2c56aeb9889c42155f6