vlserver: Always blank stack-allocated vl_ctx's
Most vlserver RPCs are structured like this pseudocode:

{
    struct vl_ctx ctx;
    if (early_error) {
        return error;
    }
    if (Init_VLdbase(&ctx, ...)) {
        return error;
    }
    if (late_error) {
        goto abort;
    }

    ...

 abort:
    vl_AbortTrans(&ctx);
    return error;
}

Note how we have some code paths that 'return' on error, and some that
'goto abort' on error. We deliberately do not 'goto abort' before
Init_VLdbase succeeds, because 'ctx' has not been initialized yet, and
may contain garbage from the stack.

While the current code doesn't make any mistakes, it's very easy to
accidentally add a 'goto abort' or otherwise access the contents of
'ctx' before 'ctx' is initialized. To avoid any such possible
accidents in the future, add a memset to clear 'ctx' for all functions
that declare a vl_ctx on the stack like this.

Change-Id: Ice0eb290690b914674ff2fbd815f9cc303ec8d03