tests: Add ubiktest multi-server support
Add support in the ubiktest framework for running tests against
multiple dbservers.

For each server process, we setup a NetInfo file to specify the single
IP 127.1.0.10, 127.1.0.11, etc, and then specify -rxbind to bind to
only that IP. Each ubiktest_dbtest is run against each site, unless
.cmd_sync is set, in which case we just run it against the sync site.

To specify which server to contact, we specify the new '-ptserver'
option to pts and the '-vlserver' option to vos.

This commit just adds the multi-server support to the ubiktest
framework; it doesn't make any tests actually use the new
functionality.

Change-Id: I8ae6b6fa00d7daf293b60978aa38b80ea9fa2b86