rx: Introduce the rx_bulk mechanism
Introduce a new mechanism to make batched or "bulk" RPC calls in rx,
called rx_bulk. Using this, a caller can call several RPCs on a
server, without needing to wait for the server response for each one.
This allows sequences of RPCs to be called more quickly, especially
for large RPC sequences, or in the presence of high network latency.

A similar strategy of issuing batched/vectorized RPCs already exists
in calls like RXAFS_InlineBulkStatus or DISK_WriteV. But all of these
involve logic and new RPCs specific for that use case (for example,
the buffering logic in ubik_Write/ubik_Flush). Introducing a mechanism
like rxbulk allows the logic for RPC batching/vectorizing to be done
consistently, and largely in the same place, reducing the need for
use-case-specific logic.

rxbulk allows for both "inline" (rxbulk_runall_inline) and
"non-inline" (rxbulk_inline) modes of operation. Using "non-inline"
mode is similar to RXAFS_BulkStatus, where a failure of any of the
given RPCs causes the whole operation to fail. Using "inline" mode is
similar to RXAFS_InlineBulkStatus, where failures are instead reported
for each inline RPC separately, and a single error does not prevent
the other RPCs from running.

To actually implement the rxbulk mechanism on the wire, the relevant
rx service needs to define an RPC for handling bulk operations
(by convention, this is called something like PKG_BulkCall). This
commit on its own doesn't define any such RPCs (except inside unit
tests); future commits will add them to rx services as callers of the
rxbulk mechanism are added.

This adds two new keywords to our .xg files processed by rxgen:
"bulk", and "bulkhandler". Specifying "bulk" after an RPC definition
(like "split" or "multi") indicates that the RPC can be specified in
rx_bulk calls (and generates the rxbulk_PKG_Foo functions). Specifying
"bulkhandler" indicates that an RPC is the RPC for interpreting
rx_bulk requests for that service.

Change-Id: I6cb601b5726930fab68c423f70f199590f405885