Introduce --enable-coverage
Introduce the --enable-coverage configure flag, to enable gcov-style
code coverage profiling. In addition to compiling and linking with
--coverage (supported by at least gcc and clang), this also defines
the ENABLE_COVERAGE symbol, to allow for some code to provide better
coverage results.

Specifically, ENABLE_COVERAGE is used to change some commands to
exit() on a fatal signal, instead of just dying to the signal. Without
this, coverage data may not be written to disk when a process dies to
a signal, and the coverage data is lost.

Change-Id: I949fcc7d02615ad5fd917400ef8e7a5e6fc180b1