rxgen: Always generate PKG_OpCodeIndex()
Currently, rxgen generates a PKG_OpCodeIndex function, to translate an
RPC's opcode into the index for that opcode (which is used for the
per-process stats, amongst other things).

However, rxgen only generates this function when there are "holes" in
the opcodes for the relevant package (when various functions are
implemented using switch statements instead of arrays). Generate this
function for the non-"holes" case as well, so callers can calculate
the index of an RPC easily for all packages.

Change-Id: I59622073cc802403bb56ac7c57dde7d282f9b1ed