rxgen: Pretend unordered opcodes have holes
Currently, rxgen generates two different variants for functions like
PKG_ExecuteRequest. If the relevant package has "holes" in the opcodes
(that is, we define op 101 and 103, but not 102), we generate a
function using a switch statement, like this:

int PKG_ExecuteRequest (struct rx_call *z_call)
{
    [...]
    switch (op) {
	case 101:
	    z_result = _PKG_Foo(z_call, &z_xdrs);
	    break;
	case 103:
	    z_result = _PKG_Baz(z_call, &z_xdrs);
	    break;
	default:
	    z_result = RXGEN_OPCODE;
	    break;
    }
    [...]
}

If the package has no opcode holes, the function is implemented using
arrays, like this:

static afs_int32 (*StubProcsArray0[])(struct rx_call *z_call, XDR *z_xdrs) = {_PKG_Foo, _PKG_Bar, _PKG_Baz};
int PKG_ExecuteRequest (struct rx_call *z_call)
{
    [...]
    if (op < PKG_LOWEST_OPCODE || op > PKG_HIGHEST_OPCODE)
	z_result = RXGEN_OPCODE;
    else
	z_result = (*StubProcsArray0[op - PKG_LOWEST_OPCODE])(z_call, &z_xdrs);
    [...]
}

However, the contents of the StubProcsArray0 are specified the order
in which the RPCs are defined in the .xg file, not the order of the
opcode numbers. So in the above example, if the order of PKG_Foo and
PKG_Bar are swapped in the defining .xg file, but the opcodes are kept
the same, the array looks like this:

static afs_int32 (*StubProcsArray0[])(struct rx_call *z_call, XDR *z_xdrs) = {_PKG_Bar, _PKG_Foo, _PKG_Baz};

This means that PKG_ExecuteRequest will execute the wrong function.
The same thing happens for PKG_TranslateOpCode, since it uses the same
array-based scheme. Currently all of the .xg files in the OpenAFS tree
without holes do have ordered opcode numbers, so this doesn't affect
any existing protocol definitions, but this is fragile and not
guaranteed to always be the case. (For instance, vldbint.xg's RPCs are
not ordered, but it currently has holes; this would be a problem if
vldbint.xg defined opcodes 525, 526, and 528-531.)

To fix this, we can just check if the opcodes for our defined RPCs are
not ordered, and act like the opcode numbers have "holes" if so.

It may be better in the future to just get rid of the array-based
implementation of PKG_ExecuteRequest et al, but for now, just work
around the issue in order to have a simple solution to the issue.

Change-Id: If5211af335629203de527d9ea9db28be3bcc03f1