rxgen: Include split calls in PKG_TranslateOpCode
Currently, in generate_code(), when 'cflag' and 'xflag' are true, we
skip adding the defp to the 'proc_defined' list if 'proc_split_flag'
is true. Normally, we want to add all procs to the 'proc_defined'
list; the reason why this is skipped for 'proc_split_flag' is
presumably because when 'cflag' is set, we call generate_code() twice
for 'multi' procs: once with proc_split_flag set, and once with it
cleared. So we skip calling generate_code() for 'proc_split_flag' to
avoid adding the proc to the global list twice.

However, for non-multi split procs, we just call generate_code once,
and 'proc_split_flag' is true. So for non-multi split procs, we'll
never add the proc to the global 'proc_defined' list (when 'cflag' is
defined). So all of that means that when generating .xdr.c files
('cflag'), non-multi split procs won't be counted in functions like
PKG_TranslateOpCode() and PKG_OpCodeIndex().

To fix this, move the STOREVAL call out of generate_code() entirely,
so we don't have to worry about the cases where generate_code() is
called twice. Just call STOREVAL one level higher, in
generate_code()'s only caller, check_proc().

This check in generate_code() was added in commit 5c77a6ca (rxgen: add
*TranslateOpcode functions to XDR code), the commit that added the
PKG_TranslateOpCode function, so this behavior has always existed for
these functions.

Change-Id: Ie83471f2560a54bc6686dc7443852d8acee1f3c3