rxgk: Introduce and use rxgk_int_error()
Add a small helper function, rxgk_int_error(), to easily log internal
or unexpected errors, so they do not happen silently. Add
rxgk_misc_error() to do the same thing for RXGK_INCONSISTENCY
specifically, to allow us to easily change the catch-all error code
used if needed.

Change almost all error paths in rxgk to use these functions. This may
seem excessive, but is very helpful if any of these rare code paths
are hit, and there's not really any downside.

A few error cases do not use these functions. Notably we do not use
these functions for expired credentials, since this case is so
uninteresting. We _do_ use these functions for errors that would be
triggered by failed attacks, since those should be very rare, and are
more likely to be bugs while rxgk is in an experimental stage.

Change-Id: Ic5b3396c4ae06161efa2714ccdc241858cf86cd4