tests: Add test rx/simple
This adds a new test, rx/simple, which runs a simple rx client against a
server processs. It does not make use of rxgen-generated RPCs, but
instead runs as a single stream of data on an rx call.

The client creates a new connection to this service and sends a string.
The server performs a simple transformation (rot13) and returns the new
string back to the client.

This commit adds the simple-client and simple-server programs, as well
as the "simple-t" script test driver. These programs serve as a very
simple example of using Rx, as well as a basic functionality test.

Written in collaboration with adeason@sinenomine.net.

Change-Id: I78862ecb75a9bb3ccbfef049d11a95182c5e0278